add product delete route and controller

+++ models/ingredient.js
@@ -10,7 +10,7 @@ module.exports = function (sequelize, DataTypes) {
             allowNull: false
         },
         quantity: {
-            type: DataTypes.INT,
+            type: DataTypes.BIGINT,
             allowNull:false
         }
     }, {
